+++ source/Core/Smarty/Plugin/function.oxinputhelp.php
@@ -5,6 +5,8 @@
  * See LICENSE file for license details.
  */
 
+use OxidEsales\Eshop\Core\Exception\StandardException;
+
 /**
  * Smarty function
  * -------------------------------------------------------------
@@ -25,7 +27,7 @@ function smarty_function_oxinputhelp($params, &$smarty)
 
     try {
         $sTranslation = $oLang->translateString($sIdent, $iLang, $blAdmin);
-    } catch (\OxidEsales\Eshop\Core\Exception\LanguageException $oEx) {
+    } catch (StandardException $oEx) {
         // is thrown in debug mode and has to be caught here, as smarty hangs otherwise!
     }