+++ source/Core/Smarty/Plugin/modifier.oxmultilangassign.php
@@ -5,6 +5,8 @@
  * See LICENSE file for license details.
  */
 
+use OxidEsales\Eshop\Core\Exception\StandardException;
+
 /**
  * Smarty function
  * -------------------------------------------------------------
@@ -36,7 +38,7 @@ function smarty_modifier_oxmultilangassign($sIdent, $args = null)
     try {
         $sTranslation = $oLang->translateString($sIdent, $iLang, $oLang->isAdmin());
         $blTranslationNotFound = !$oLang->isTranslated();
-    } catch (\OxidEsales\Eshop\Core\Exception\LanguageException $oEx) {
+    } catch (StandardException $oEx) {
         // is thrown in debug mode and has to be caught here, as smarty hangs otherwise!
     }