+++ source/Core/Smarty/Plugin/modifier.oxmultilangsal.php
@@ -5,6 +5,8 @@
  * See LICENSE file for license details.
  */
 
+use OxidEsales\Eshop\Core\Exception\StandardException;
+
 /**
  * Smarty function
  * -------------------------------------------------------------
@@ -30,7 +32,7 @@ function smarty_modifier_oxmultilangsal($sIdent)
 
     try {
         $sTranslation = $oLang->translateString($sIdent, $iLang, $oLang->isAdmin());
-    } catch (\OxidEsales\Eshop\Core\Exception\LanguageException $oEx) {
+    } catch (StandardException $oEx) {
         // is thrown in debug mode and has to be caught here, as smarty hangs otherwise!
     }