+++ source/Internal/Transition/Adapter/TemplateLogic/TranslateFilterLogic.php
@@ -7,6 +7,8 @@
 
 namespace OxidEsales\EshopCommunity\Internal\Transition\Adapter\TemplateLogic;
 
+use OxidEsales\Eshop\Core\Exception\StandardException;
+
 class TranslateFilterLogic
 {
     /**
@@ -34,7 +36,7 @@ class TranslateFilterLogic
         try {
             $sTranslation = $oLang->translateString($sIdent, $iLang, $oLang->isAdmin());
             $blTranslationNotFound = !$oLang->isTranslated();
-        } catch (\OxidEsales\Eshop\Core\Exception\LanguageException $oEx) {
+        } catch (StandardException $oEx) {
             // is thrown in debug mode and has to be caught here, as smarty hangs otherwise!
         }