+++ CHANGELOG.md
@@ -329,6 +329,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `\OxidEsales\EshopCommunity\Core\WidgetControl::$_blHandlerSet`
 - `\OxidEsales\EshopCommunity\Core\ViewConfig::getSession()`
 - `\OxidEsales\EshopCommunity\Core\Routing\Module\ClassProviderStorage`
+- `\OxidEsales\EshopCommunity\Core\Exception\LanguageException`
 
 ### Fixed
 - Fix not working actions and promotions [#0005526](https://bugs.oxid-esales.com/view.php?id=5526)